django/11일차(25.11.15)/practice/example/serializers.py
from rest_framework import serializers
from .models import Books

class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Books
        fields = [
            'bid',
            'title',
            'author',
            'category',
            'pages',
            'price',
            'published_date',
            'description'
        ]